Fixes for Windows
The code isn't working yet, but it's closer.

ipc-win.cpp:
+ set thread names

renderer-backend.cpp:
+ remove commented out code.
+ do not call eglCreateWindowSurface.  EGLTarget::initialize is
  expected to create the native window; creating an EGL display
  surface is done later by WebKit.
+ move code from display.cpp for handling input messages.

display.cpp:
+ remove code for handling Windows messages.

display.h:
+ remove commented code.

pipe.cpp:
+ for a server, call WaitForClient in ReceiveThread.
+ start ReceiveThread automatically for clients and servers.
